Moved validators to validators file
